requirements.txt
streamlit-navigation-bar==3.3.0
st-circular-progress==0.3.9